+++ src/infrastructure/mod.rs
@@ -2,7 +2,11 @@
 pub mod file_system;
 pub mod processors;
 pub mod ultra_file_system;
+pub mod hmr;
+pub mod hmr_client;
 
 pub use file_system::*;
 pub use processors::*;
 pub use ultra_file_system::*;
+pub use hmr::*;
+pub use hmr_client::*;